Views Fix / Can't edit profiles that isn't yours / Profile and profile edit views

+++ routes/web.php
@@ -6,6 +6,7 @@ use App\Http\Controllers\HomeController;
 use App\Http\Controllers\UserController;
 use App\Http\Controllers\GameController;
 use App\Http\Controllers\PanelController;
+use App\Http\Controllers\PerfilController;
 use App\Http\Controllers\ReviewController;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Redirect;
@@ -24,9 +25,16 @@ use Illuminate\Support\Facades\Redirect;
 Auth::routes();
 
 Route::get('/', [HomeController::class, 'index'])->name('home');
-Route::get('/games', [GameController::class,'index'])->name('games');
+
+Route::get('/games', [GameController::class,'index'])->name('games.index');
 Route::get('/games/search', [GameController::class,'search'])->name('games.search');
-Route::get('/panel', [PanelController::class, 'index'])->name('panel')->middleware("admin");
+
+Route::get('/perfil/{perfil}', [PerfilController::class,'show'])->name('perfil.index');
+Route::get('/perfil/{perfil}/edit', [PerfilController::class,'edit'])->name('perfilEdit')->middleware("userprofile");
+Route::get('/perfil/{perfil}/update', [PerfilController::class,'update'])->name('perfil.update');
+
+Route::get('/panel', [PanelController::class, 'index'])->name('panel.index')->middleware("admin");
+
 Route::get('/logout', function ()
 {
     auth()->logout();
@@ -37,37 +45,37 @@ Route::get('/logout', function ()
 
 
 //CRUD USER
-Route::get('panel/users/index', [UserController::class,'index'])->name('users.index');
-Route::get('panel/users/create', [UserController::class,'create'])->name('users.create');
-Route::delete('panel/users/destroy/{user}', [UserController::class,'destroy'])->name('users.destroy');
-Route::get('panel/users/show/{user}', [UserController::class,'show'])->name('users.show');
-Route::get('panel/users/edit/{user}', [UserController::class,'edit'])->name('users.edit');
-Route::put('panel/users/update/{user}', [UserController::class,'update'])->name('users.update');
-Route::post('panel/users/store', [UserController::class,'store'])->name('users.store');
+Route::get('panel/users', [UserController::class,'index'])->name('users.index')->middleware("admin");
+Route::get('panel/users/create', [UserController::class,'create'])->name('users.create')->middleware("admin");
+Route::get('panel/users/{user}', [UserController::class,'show'])->name('users.show')->middleware("admin");
+Route::get('panel/users/{user}/edit', [UserController::class,'edit'])->name('users.edit')->middleware("admin");
+Route::post('panel/users', [UserController::class,'store'])->name('users.store')->middleware("admin");
+Route::put('panel/users/{user}', [UserController::class,'update'])->name('users.update')->middleware("admin");
+Route::delete('panel/users/{user}/destroy', [UserController::class,'destroy'])->name('users.destroy')->middleware("admin");
 
 //CRUD GAME
-Route::get('panel/games/resultados', [GameController::class,'resultados'])->name('games.resultados');
-Route::get('panel/games/create', [GameController::class,'create'])->name('games.create');
-Route::delete('panel/games/destroy/{game}', [GameController::class,'destroy'])->name('games.destroy');
-Route::get('panel/games/show/{game}', [GameController::class,'show'])->name('games.show');
-Route::get('panel/games/edit/{game}', [GameController::class,'edit'])->name('games.edit');
-Route::put('panel/games/update/{game}', [GameController::class,'update'])->name('games.update');
-Route::post('panel/games/store', [GameController::class,'store'])->name('games.store');
+Route::get('panel/games', [GameController::class,'panelIndex'])->name('games.panelIndex')->middleware("admin");
+Route::get('panel/games/create', [GameController::class,'create'])->name('games.create')->middleware("admin");
+Route::delete('panel/games/destroy/{game}', [GameController::class,'destroy'])->name('games.destroy')->middleware("admin");
+Route::get('panel/games/show/{game}', [GameController::class,'show'])->name('games.show')->middleware("admin");
+Route::get('panel/games/edit/{game}', [GameController::class,'edit'])->name('games.edit')->middleware("admin");
+Route::put('panel/games/update/{game}', [GameController::class,'update'])->name('games.update')->middleware("admin");
+Route::post('panel/games/store', [GameController::class,'store'])->name('games.store')->middleware("admin");
 
 //CRUD REVIEW
-Route::get('panel/reviews/index', [ReviewController::class,'index'])->name('reviews.index');
-Route::get('panel/reviews/create', [ReviewController::class,'create'])->name('reviews.create');
-Route::delete('panel/reviews/destroy/{review}', [ReviewController::class,'destroy'])->name('reviews.destroy');
-Route::get('panel/reviews/show/{review}', [ReviewController::class,'show'])->name('reviews.show');
-Route::get('panel/reviews/edit/{review}', [ReviewController::class,'edit'])->name('reviews.edit');
-Route::put('panel/reviews/update/{review}', [ReviewController::class,'update'])->name('reviews.update');
-Route::post('panel/reviews/store', [ReviewController::class,'store'])->name('reviews.store');
+Route::get('panel/reviews/index', [ReviewController::class,'index'])->name('reviews.index')->middleware("admin");
+Route::get('panel/reviews/create', [ReviewController::class,'create'])->name('reviews.create')->middleware("admin");
+Route::delete('panel/reviews/destroy/{review}', [ReviewController::class,'destroy'])->name('reviews.destroy')->middleware("admin");
+Route::get('panel/reviews/show/{review}', [ReviewController::class,'show'])->name('reviews.show')->middleware("admin");
+Route::get('panel/reviews/edit/{review}', [ReviewController::class,'edit'])->name('reviews.edit')->middleware("admin");
+Route::put('panel/reviews/update/{review}', [ReviewController::class,'update'])->name('reviews.update')->middleware("admin");
+Route::post('panel/reviews/store', [ReviewController::class,'store'])->name('reviews.store')->middleware("admin");
 
 //CRUD COMMENT
-Route::get('panel/comments/index', [CommentController::class,'index'])->name('comments.index');
-Route::get('panel/comments/create', [CommentController::class,'create'])->name('comments.create');
-Route::delete('panel/comments/destroy/{comment}', [CommentController::class,'destroy'])->name('comments.destroy');
-Route::get('panel/comments/show/{comment}', [CommentController::class,'show'])->name('comments.show');
-Route::get('panel/comments/edit/{comment}', [CommentController::class,'edit'])->name('comments.edit');
-Route::put('panel/comments/update/{comment}', [CommentController::class,'update'])->name('comments.update');
-Route::post('panel/comments/store', [CommentController::class,'store'])->name('comments.store');
+Route::get('panel/comments/index', [CommentController::class,'index'])->name('comments.index')->middleware("admin");
+Route::get('panel/comments/create', [CommentController::class,'create'])->name('comments.create')->middleware("admin");
+Route::delete('panel/comments/destroy/{comment}', [CommentController::class,'destroy'])->name('comments.destroy')->middleware("admin");
+Route::get('panel/comments/show/{comment}', [CommentController::class,'show'])->name('comments.show')->middleware("admin");
+Route::get('panel/comments/edit/{comment}', [CommentController::class,'edit'])->name('comments.edit')->middleware("admin");
+Route::put('panel/comments/update/{comment}', [CommentController::class,'update'])->name('comments.update')->middleware("admin");
+Route::post('panel/comments/store', [CommentController::class,'store'])->name('comments.store')->middleware("admin");
